+++ tests/integration_tests/reference_files/celery_sso.cloudron_settings.py
@@ -15,6 +15,8 @@ if os.environ.get("CLOUDRON_APP_ORIGIN"):
     SECURE_PROXY_SSL_HEADER = ("HTTP_X_FORWARDED_PROTO", "https")
     SESSION_COOKIE_SECURE = True
     CSRF_COOKIE_SECURE = True
+    SECURE_CONTENT_TYPE_NOSNIFF = True
+    SECURE_REFERRER_POLICY = "same-origin"
 
     DATABASES = {
         "default": {